refactor <Grid> sample and default configs
